+++ server/src/index.ts
@@ -28,6 +28,7 @@ import "./modules/routine/resolvers/routineTriggers";
 import "./modules/attendance/resolvers/teacherAttendance";
 import "./modules/attendance/resolvers/studentAttendance";
 import "./modules/attendance/resolvers/push";
+import "./modules/hr/resolvers/staffLeave";
 import "./modules/guardian/resolvers/guardianPortal";
 import "./modules/notifications/resolvers/notifications";
 import "./modules/library/resolvers/library";